CTJoT: Fix for immediate finish/release on connect bug
Sometimes the game would mark as complete immediately on connect, even
with a fresh, umplayed game.  Current assumption is that this is caused
by an order of operations issue, where the game is started after the
client, and the hardware/emulator is filling RAM with some pre-defined,
non-zero value that is being interpreted incorrectly by the client as
the game completion flag being set.

This new code checks the location value to ensure that the current map
is a valid value before continuing with tracking.  If the map is
invalid, bail out and don't continue the tracking function.